# Parameterized abstractions that work on CPU and GPU architectures.