--- a/src/boundary_conditions.jl
+++ b/src/boundary_conditions.jl
@@ -106,3 +106,85 @@ end
 
 const BC = BoundaryCondition
 const FBCs = FieldBoundaryConditions
+
+#
+# Physics goes here.
+#
+
+#=
+Currently we support flux and gradient boundary conditions
+at the top and bottom of the domain.
+
+Notes:
+
+- The boundary condition on a z-boundary is a callable object with arguments
+
+      (t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S, iteration, i, j),
+
+  where i and j are the x and y indices, respectively. No other function signature will work.
+  We do not have abstractions that generalize to non-uniform grids.
+
+- We assume that the boundary tendency has been previously calculated for
+  a 'no-flux' boundary condition.
+
+  This means that boudnary conditions take the form of
+  an addition/subtraction to the tendency associated with a flux at point (A, A, I) below the bottom cell.
+  This paradigm holds as long as consider boundary conditions on (A, A, C) variables only, where A is
+  "any" of C or I.
+
+ - We use the physics-based convention that
+
+        flux = -κ * gradient,
+
+    and that
+
+        tendency = ∂ϕ/∂t = Gϕ = - ∇ ⋅ flux
+
+=#
+
+# Do nothing in default case. These functions are called in cases where one of the
+# z-boundaries is set, but not the other.
+@inline apply_z_top_bc!(args...) = nothing
+@inline apply_z_bottom_bc!(args...) = nothing
+
+# These functions compute vertical fluxes for (A, A, C) quantities. They are not currently used.
+@inline ∇κ∇ϕ_t(κ, ϕt, ϕt₋₁, flux, Δzc, Δzf) = (      -flux        - κ*(ϕt - ϕt₋₁)/Δzc ) / Δzf
+@inline ∇κ∇ϕ_b(κ, ϕb, ϕb₊₁, flux, Δzc, Δzf) = ( κ*(ϕb₊₁ - ϕb)/Δzc +       flux        ) / Δzf
+
+"Add flux divergence to ∂ϕ/∂t associated with a top boundary condition on ϕ."
+@inline apply_z_top_bc!(top_flux::BC{<:Flux, <:Function}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, 1] -= top_flux.condition(t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S, iteration, i, j) / Δz
+
+@inline apply_z_top_bc!(top_flux::BC{<:Flux, <:Number}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, 1] -= top_flux.condition / Δz
+
+@inline apply_z_top_bc!(top_flux::BC{<:Flux, <:AbstractArray}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, 1] -= top_flux.condition[i, j] / Δz
+
+@inline apply_z_top_bc!(top_gradient::BC{<:Gradient, <:Function}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, 1] += κ*top_gradient.condition(t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S, iteration, i, j) / Δz
+
+@inline apply_z_top_bc!(top_gradient::BC{<:Gradient, <:Number}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, 1] += κ*top_gradient.condition / Δz
+
+@inline apply_z_top_bc!(top_gradient::BC{<:Gradient, <:AbstractArray}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, 1] += κ*top_gradient.condition[i, j] / Δz
+
+"Add flux divergence to ∂ϕ/∂t associated with a bottom boundary condition on ϕ."
+@inline apply_z_bottom_bc!(bottom_flux::BC{<:Flux, <:Function}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, Nz] += bottom_flux.condition(t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S, iteration, i, j) / Δz
+
+@inline apply_z_bottom_bc!(bottom_flux::BC{<:Flux, <:Number}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, Nz] += bottom_flux.condition / Δz
+
+@inline apply_z_bottom_bc!(bottom_flux::BC{<:Flux, <:AbstractArray}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, Nz] += bottom_flux.condition[i, j] / Δz
+
+@inline apply_z_bottom_bc!(bottom_gradient::BC{<:Gradient, <:Function}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, Nz] -= κ*bottom_gradient.condition(t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S, iteration, i, j) / Δz
+
+@inline apply_z_bottom_bc!(bottom_gradient::BC{<:Gradient, <:Number}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, Nz] -= κ*bottom_gradient.condition / Δz
+
+@inline apply_z_bottom_bc!(bottom_gradient::BC{<:Gradient, <:AbstractArray}, ϕ, Gϕ, κ, t, Δx, Δy, Δz, Nx, Ny, Nz, u, v, w, T, S,
+    iteration, i, j) = Gϕ[i, j, Nz] -= κ*bottom_gradient.condition[i, j] / Δz