This PR refactors many of the existing abstractions so that they can be passed to GPU kernels #59. It massively cleans up the operators and time-stepping loop, allowing us to keep using beautiful abstractions even in GPU kernels.

Note that after this PR is merged, `CUDAnative#master` and `GPUifyLoops#vc/interactive` will be required until the respective fixes are merged into an updated release. I will update the `Project.toml` when we can move to a new release.

Once this is merged I will also release v0.5.0.

When completed and merged this PR will:
Resolve #19
Resolve #59
Resolve #60
Resolve #66
Resolve #74 
Resolve #133 
Resolve #153