@ali-ramadhan 2019-03-26
Thanks for having a look! I'm starting to clean up the time stepping now.

Will definitely request a review from you once this is done so we can iterate on something that works for everyone.