@ali-ramadhan 2019-04-03
I could keep cleaning things up but I think I've done enough to close #59.

The only big feature missing is turning our `Field` struct into something we can adapt/convert to a CUDA device argument using `cudaconvert` or Adapt.jl. We can do more cleanup when we figure out how to do this.

Will release v0.5.0 once this is merged.