@ali-ramadhan 2019-04-03
So I just ran the performance benchmarks again (different CPU but same GPU) and the model has gotten slower since the last time we ran benchmarks (PR #116), by about 30~40%.

Details pasted below but for the larger GPU models: 256^3 Float32 it went up from 85.9 ms to 119 ms, and for 128^3 Float32 it went up from 7.70 ms to 10.1 ms. These are wall clock time per model iteration.

To be fair, we had no boundary conditions API last time we benchmarked. I won't worry about this for now though as we haven't begun to optimize for performance.

```
──────────────────────────────────────────────────────────────────────────────────────────────────
             Oceananigans.jl benchmarks                    Time                   Allocations      
                                                   ──────────────────────   ───────────────────────
                 Tot / % measured:                       734s / 72.9%           18.7GiB / 0.06%    

 Section                                   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────────────────────────
 256x256x256 static ocean (CPU, Float32)       10     312s  58.4%   31.2s   60.2KiB  0.47%  6.02KiB
 256x256x256 static ocean (CPU, Float64)       10     196s  36.7%   19.6s   78.0KiB  0.62%  7.80KiB
 128x128x128 static ocean (CPU, Float32)       10    14.2s  2.66%   1.42s   60.2KiB  0.47%  6.02KiB
 128x128x128 static ocean (CPU, Float64)       10    7.64s  1.43%   764ms   78.0KiB  0.62%  7.80KiB
 256x256x256 static ocean (GPU, Float64)       10    1.36s  0.25%   136ms   1.60MiB  12.9%   164KiB
 256x256x256 static ocean (GPU, Float32)       10    1.19s  0.22%   119ms   1.36MiB  11.0%   139KiB
  64x 64x 64 static ocean (CPU, Float32)       10    950ms  0.18%  95.0ms   60.2KiB  0.47%  6.02KiB
  64x 64x 64 static ocean (CPU, Float64)       10    584ms  0.11%  58.4ms   78.0KiB  0.62%  7.80KiB
 128x128x128 static ocean (GPU, Float64)       10    116ms  0.02%  11.6ms   1.60MiB  12.9%   164KiB
 128x128x128 static ocean (GPU, Float32)       10    101ms  0.02%  10.1ms   1.36MiB  11.0%   139KiB
  32x 32x 32 static ocean (CPU, Float32)       10   94.6ms  0.02%  9.46ms   60.2KiB  0.47%  6.02KiB
  32x 32x 32 static ocean (CPU, Float64)       10   53.9ms  0.01%  5.39ms   78.0KiB  0.62%  7.80KiB
  64x 64x 64 static ocean (GPU, Float64)       10   11.6ms  0.00%  1.16ms   1.60MiB  12.9%   164KiB
  64x 64x 64 static ocean (GPU, Float32)       10   11.2ms  0.00%  1.12ms   1.36MiB  11.0%   139KiB
  32x 32x 32 static ocean (GPU, Float32)       10   5.61ms  0.00%   561μs   1.36MiB  11.0%   139KiB
  32x 32x 32 static ocean (GPU, Float64)       10   5.18ms  0.00%   518μs   1.60MiB  12.9%   164KiB
 ──────────────────────────────────────────────────────────────────────────────────────────────────

CPU Float64 -> Float32 speedup:
 32x 32x 32 static ocean: 0.570
 64x 64x 64 static ocean: 0.615
128x128x128 static ocean: 0.538
256x256x256 static ocean: 0.628

GPU Float64 -> Float32 speedup:
 32x 32x 32 static ocean: 0.923
 64x 64x 64 static ocean: 1.038
128x128x128 static ocean: 1.147
256x256x256 static ocean: 1.138

CPU -> GPU speedup:
 32x 32x 32 static ocean (Float32): 16.867
 32x 32x 32 static ocean (Float64): 10.414
 64x 64x 64 static ocean (Float32): 84.867
 64x 64x 64 static ocean (Float64): 50.283
128x128x128 static ocean (Float32): 139.956
128x128x128 static ocean (Float64): 65.665
256x256x256 static ocean (Float32): 261.842
256x256x256 static ocean (Float64): 144.460
```