@vchuravy 2019-04-03
> Note that tests will fail on dev/nightly builds (Julia 1.2) because something changed which broke Cassette (which GPUifyLoops depends on).

Pushed a fix for that this morning, and  also tagged GPUifyLoops v0.2.1