@glwagner 2019-04-04
Hmm... if the boundary conditions are default, the only thing that happens is a function is called which is something like `f(args...) = nothing`. That might get compiled away too (I was hoping so). So there's a bug in the boundary condition implementation if its causing a slowdown with default boundary conditions. 30-40% is too much slowdown for adding boundary conditions anyways I think.