@ali-ramadhan 2019-04-04
That's true, there's been a ton of refactoring so it's probably more than just one thing that's changed (I can't say for sure that my changes didn't slow things down).

No point worrying too much until we get around to profiling #162.